Bluetooth layer blueprint
